refactor: centralize stats calculation logic into a utility function and simplify API handlers
